renamed to modal reducer
