+++ frontend/src/reducers/modal_reducer.js
@@ -1,4 +1,4 @@
-import { OPEN_MODAL, CLOSE_MODAL } from "../actions/user_modal_actions";
+import { OPEN_MODAL, CLOSE_MODAL } from "../actions/modal_actions";
 
 export default function modalReducer(state = null, action) {
   switch (action.type) {
