+++ frontend/src/reducers/ui_reducer.js
@@ -1,6 +1,6 @@
 import { combineReducers } from "redux";
 
-import modalReducer from "./user_modal_reducer";
+import modalReducer from "./modal_reducer";
 
 const uiReducer = combineReducers({
   modal: modalReducer,
